Modify: Split audio player onto dependent file
